+++ packages/time_config_checker/time_config_checker/CHANGELOG.md
@@ -1,3 +1,47 @@
+## 0.0.12
+
+* Fix: Import content resolver
+
+## 0.0.11
+
+* Fix: Import content resolver
+
+## 0.0.10
+
+* Fix: Import content resolver
+
+## 0.0.9
+
+* Fix: Import content resolver
+
+## 0.0.8
+
+* Testing
+
+## 0.0.7
+
+* Testing
+
+## 0.0.6
+
+* Testing
+
+## 0.0.5
+
+* Testing
+
+## 0.0.4
+
+* Initial testing
+
+## 0.0.3
+
+* Initial testing
+
+## 0.0.2
+
+* Initial testing
+
 ## 0.0.1
 
 * Test release