+++ packages/time_config_checker/time_config_checker_platform_interface/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 0.0.6
+
+* Testing
+
+## 0.0.5
+
+* Testing
+
+## 0.0.4
+
+* Initial testing
+
+## 0.0.3
+
+* Initial testing
+
+## 0.0.2
+
+* Initial testing
+
 ## 0.0.1
 
 * Test release